# openapi.yaml
openapi: 3.0.3
info:
  title: CubeOS API
  description: |
    CubeOS unified API for managing Docker services, network configuration,
    VPN connections, and storage mounts on Raspberry Pi.
  version: 2.0.0
  contact:
    name: CubeOS
    url: https://cubeos.app
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://api.cubeos.cube/api/v1
    description: Internal CubeOS network
  - url: http://10.42.24.1:6010/api/v1
    description: Direct IP access
...
security:
  - bearerAuth: []
paths:
...
  /vpn/configs:
    get:
      tags: [VPN]
      summary: List VPN configurations
      responses:
        '200':
          description: List of VPN configs
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      $ref: '#/components/schemas/VPNConfig'
    post:
      tags: [VPN]
      summary: Add VPN configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddVPNConfigRequest'
      responses:
        '201':
          description: VPN config added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VPNConfig'
  /vpn/configs/{name}:
    get:
      tags: [VPN]
      summary: Get VPN configuration
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VPN config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VPNConfig'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [VPN]
      summary: Delete VPN configuration
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VPN config deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /vpn/configs/{name}/connect:
    post:
      tags: [VPN]
      summary: Connect VPN
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VPN connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  details:
                    $ref: '#/components/schemas/VPNStatus'
        '500':
          $ref: '#/components/responses/InternalError'
  /vpn/configs/{name}/disconnect:
    post:
      tags: [VPN]
      summary: Disconnect VPN
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VPN disconnected
  /vpn/configs/{name}/auto-connect:
    put:
      tags: [VPN]
      summary: Set VPN auto-connect
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Auto-connect updated
...
  /mounts:
    get:
      tags: [Mounts]
      summary: List all mounts
      responses:
        '200':
          description: List of mounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  mounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mount'
    post:
      tags: [Mounts]
      summary: Add mount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMountRequest'
      responses:
        '201':
          description: Mount added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mount'
  /mounts/test:
    post:
      tags: [Mounts]
      summary: Test mount connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, remote_path]
              properties:
                type:
                  type: string
                  enum: [smb, nfs]
                remote_path:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
  /mounts/{name}:
    get:
      tags: [Mounts]
      summary: Get mount details
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mount details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mount'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Mounts]
      summary: Delete mount
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mount deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /mounts/{name}/mount:
    post:
      tags: [Mounts]
      summary: Mount share
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share mounted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  details:
                    $ref: '#/components/schemas/MountStatus'
  /mounts/{name}/unmount:
    post:
      tags: [Mounts]
      summary: Unmount share
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share unmounted
  /mounts/{name}/status:
    get:
      tags: [Mounts]
      summary: Get mount status
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mount status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountStatus'

  # ============================================================
  # System
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    VPNStatus:
      type: object
      properties:
        connected:
          type: boolean
        active_config:
          type: string
        type:
          type: string
          enum: [wireguard, openvpn]
        public_ip:
          type: string
        connected_at:
          type: string
          format: date-time
        bytes_sent:
          type: integer
        bytes_received:
          type: integer
    VPNConfig:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [wireguard, openvpn]
        is_active:
          type: boolean
        auto_connect:
          type: boolean
        created_at:
          type: string
          format: date-time
    AddVPNConfigRequest:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [wireguard, openvpn]
        config:
          type: string
          description: Base64 encoded config file
    Mount:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [smb, nfs]
        remote_path:
          type: string
        local_path:
          type: string
        username:
          type: string
        is_mounted:
          type: boolean
        auto_mount:
          type: boolean
        created_at:
          type: string
          format: date-time
    AddMountRequest:
      type: object
      required: [name, type, remote_path]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [smb, nfs]
        remote_path:
          type: string
          description: "SMB: //server/share, NFS: server:/path"
        username:
          type: string
        password:
          type: string
        options:
          type: string
        auto_mount:
          type: boolean
    MountStatus:
      type: object
      properties:
        name:
          type: string
        is_mounted:
          type: boolean
        local_path:
          type: string
        remote_path:
          type: string
        type:
          type: string
        used_bytes:
          type: integer
        total_bytes:
          type: integer
        free_bytes:
          type: integer
...